[JIT] Reuse traced fn for jit opinfos
Previously, jit opinfos would only run the traced function once. This is a problem for NNC and NVFuser, where the fused implementation only runs on the second invocation.

This caches the traced function and calls the cached implementation, so that subsequent calls actually perform fusion and use the fused implementation.

ghstack-source-id: 582124267855398049dc5f402cb5a2ed56b5164d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/76000